pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jean-claude"
version = "0.1.0"
description = "CLI foundation for the Jean-Claude machine assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
]

[project.scripts]
jc = "jean_claude.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
